Move integer type definition out of convert into types module
