+++ artichoke-backend/src/convert/array/fixnum.rs
@@ -1,6 +1,5 @@
-use crate::convert::fixnum::Int;
 use crate::convert::{Convert, Error, TryConvert};
-use crate::types::{Ruby, Rust};
+use crate::types::{Int, Ruby, Rust};
 use crate::value::Value;
 use crate::Artichoke;
 
@@ -73,11 +72,10 @@ mod tests {
     use quickcheck_macros::quickcheck;
     use std::convert::TryFrom;
 
-    use crate::convert::fixnum::Int;
     use crate::convert::{Convert, Error, TryConvert};
     use crate::eval::Eval;
     use crate::sys;
-    use crate::types::{Ruby, Rust};
+    use crate::types::{Int, Ruby, Rust};
     use crate::value::Value;
 
     #[test]
