+++ artichoke-backend/src/convert/nilable/array.rs
@@ -1,7 +1,6 @@
-use crate::convert::fixnum::Int;
 use crate::convert::float::Float;
 use crate::convert::{Convert, Error, TryConvert};
-use crate::types::{Ruby, Rust};
+use crate::types::{Int, Ruby, Rust};
 use crate::value::Value;
 use crate::Artichoke;
 
@@ -182,11 +181,10 @@ impl Convert<Option<Vec<&str>>> for Value {
 mod tests {
     use quickcheck_macros::quickcheck;
 
-    use crate::convert::fixnum::Int;
     use crate::convert::{Convert, TryConvert};
     use crate::eval::Eval;
     use crate::sys;
-    use crate::types::Ruby;
+    use crate::types::{Int, Ruby};
     use crate::value::Value;
 
     #[test]
