+++ artichoke-backend/src/extn/core/matchdata/length.rs
@@ -2,8 +2,9 @@
 
 use std::convert::TryFrom;
 
-use crate::convert::{Convert, Int, RustBackedValue};
+use crate::convert::{Convert, RustBackedValue};
 use crate::extn::core::matchdata::MatchData;
+use crate::types::Int;
 use crate::value::Value;
 use crate::Artichoke;
