+++ artichoke-backend/src/extn/core/regexp/fixed_encoding.rs
@@ -2,9 +2,10 @@
 
 use std::convert::TryFrom;
 
-use crate::convert::{Convert, Int, RustBackedValue};
+use crate::convert::{Convert, RustBackedValue};
 use crate::extn::core::regexp::enc::Encoding;
 use crate::extn::core::regexp::Regexp;
+use crate::types::Int;
 use crate::value::Value;
 use crate::Artichoke;
