+++ artichoke-backend/src/extn/core/regexp/hash.rs
@@ -3,8 +3,9 @@
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 
-use crate::convert::{Convert, Int, RustBackedValue};
+use crate::convert::{Convert, RustBackedValue};
 use crate::extn::core::regexp::Regexp;
+use crate::types::Int;
 use crate::value::Value;
 use crate::Artichoke;
