+++ artichoke-backend/src/extn/core/regexp/match_operator.rs
@@ -4,10 +4,11 @@ use std::cmp;
 use std::convert::TryFrom;
 use std::mem;
 
-use crate::convert::{Convert, Int, RustBackedValue, TryConvert};
+use crate::convert::{Convert, RustBackedValue, TryConvert};
 use crate::extn::core::matchdata::MatchData;
 use crate::extn::core::regexp::Regexp;
 use crate::sys;
+use crate::types::Int;
 use crate::value::Value;
 use crate::Artichoke;
